requirements.txt
Flask==2.0.3
iniconfig==1.1.1
packaging==21.3
pandas==1.1.5
py==1.11.0
pyodbc==4.0.39
pyparsing==3.1.2
pytest==7.0.1
SQLAlchemy==1.4.52
python-dotenv==0.20.0